Improve user friendship collection
After every collected user, a separated file is created as backup which allows to work with the user friendships in the process. The user friendship would be otherwise overwritten every other second. Additionally the progress-points were deleted as the file user_friendships_secondary contains the information from Baerbock and Scholz as well.